feat(sourcemaps): add SentryFacade and Sentry CLI Boilerplate (#23)
This PR marks the first step towards getting source maps and release creation/finalization working. It adds Sentry-CLI as a dependency, as well as a facade object for the necessary steps (see explanation below).

While we want to get rid of depending on Sentry CLI eventually, I decided to use it for now to get our MVP out faster. To replace Sentry CLI functionality easily with our "native" functionality later, I created a facade to abstract the implementation of the necessary steps from what we call in the "front end" of our plugin. We can replace each CLI command step by step with our own implementation later on.