fix: Fix regex in source map locating heuristic via `sourceMappingURL`
Fixes #375

We have a regex that tries to find the path of a source map of a file via the `//# sourceMappingURL=` comment. It never worked because the this regex was intended to be a multiline regex ... which it wasn't.

This PR turns it into a multiline regex. While we're at it, let's also allow some whitespace in front of the comment.