feat(component-annotate): Introduce new plugin to annotate frontend components at build-time (#468)
This PR in particular adds a new package to the bundler plugins; a Babel plugin that will annotate frontend components at build-time with additional data. A lot of the source code is vendored from FullStory's React annotate plugin, but has been modified and converted to TypeScript. For now, the plugin only works on React, but we can extend it to other frontend frameworks / libraries in the future.

---------

Co-authored-by: George Gritsouk <989898+gggritso@users.noreply.github.com>
Co-authored-by: Abhijeet Prasad <aprasad@sentry.io>