fix(core): Observe and handle Sentry CLI sourcemap upload failures (#763)
Bump `@sentry/cli` to 2.49.0 which adds a new `rejectOnError` live mode setting to the `execute` and `releases.uploadSourceMaps` APIs. 

This new mode now makes the CLI reject the promise it returns if the CLI binary process exits with a non-zero exit code. In the previous live mode, it never rejected but only piped the stdio output, meaning the plugins didn't even know that the CLI failed.