fix(v2/core): Make `moduleMetadata` injection code ES5-compliant (#773)
Makes the v2 `moduleMetadata` code snippet ES5-compliant by replacing the `Object.assign` merge strategy of the module with a pure ES5 merge strategy: iterating over the object property of both objects to merge to the final object.
Obviously, this increases the bundle size footprint but I think we need to prioritize correctness.

Also took the opportunity to minify the code and further isolate the `var`-declared variables by wrapping the code in an IIFE. Adjusted to the tests to also use `toMatchSnapshot`, since the inline snapshots would no longer be readable.